+++ src/backend/core/tests/documents/test_api_documents_retrieve.py
@@ -28,8 +28,7 @@ def test_api_documents_retrieve_anonymous_public_standalone():
         "abilities": {
             "accesses_manage": False,
             "accesses_view": False,
-            "ai_transform": False,
-            "ai_translate": False,
+            "ai_proxy": False,
             "attachment_upload": document.link_role == "editor",
             "children_create": False,
             "children_list": True,
@@ -96,8 +95,7 @@ def test_api_documents_retrieve_anonymous_public_parent():
         "abilities": {
             "accesses_manage": False,
             "accesses_view": False,
-            "ai_transform": False,
-            "ai_translate": False,
+            "ai_proxy": False,
             "attachment_upload": grand_parent.link_role == "editor",
             "children_create": False,
             "children_list": True,
@@ -193,8 +191,7 @@ def test_api_documents_retrieve_authenticated_unrelated_public_or_authenticated(
         "abilities": {
             "accesses_manage": False,
             "accesses_view": False,
-            "ai_transform": document.link_role == "editor",
-            "ai_translate": document.link_role == "editor",
+            "ai_proxy": document.link_role == "editor",
             "attachment_upload": document.link_role == "editor",
             "children_create": document.link_role == "editor",
             "children_list": True,
@@ -268,8 +265,7 @@ def test_api_documents_retrieve_authenticated_public_or_authenticated_parent(rea
         "abilities": {
             "accesses_manage": False,
             "accesses_view": False,
-            "ai_transform": grand_parent.link_role == "editor",
-            "ai_translate": grand_parent.link_role == "editor",
+            "ai_proxy": grand_parent.link_role == "editor",
             "attachment_upload": grand_parent.link_role == "editor",
             "children_create": grand_parent.link_role == "editor",
             "children_list": True,
@@ -449,8 +445,7 @@ def test_api_documents_retrieve_authenticated_related_parent():
         "abilities": {
             "accesses_manage": access.role in ["administrator", "owner"],
             "accesses_view": True,
-            "ai_transform": access.role != "reader",
-            "ai_translate": access.role != "reader",
+            "ai_proxy": access.role != "reader",
             "attachment_upload": access.role != "reader",
             "children_create": access.role != "reader",
             "children_list": True,