+++ src/backend/core/tests/test_models_documents.py
@@ -152,8 +152,7 @@ def test_models_documents_get_abilities_forbidden(
     expected_abilities = {
         "accesses_manage": False,
         "accesses_view": False,
-        "ai_transform": False,
-        "ai_translate": False,
+        "ai_proxy": False,
         "attachment_upload": False,
         "children_create": False,
         "children_list": False,
@@ -213,8 +212,7 @@ def test_models_documents_get_abilities_reader(
     expected_abilities = {
         "accesses_manage": False,
         "accesses_view": False,
-        "ai_transform": False,
-        "ai_translate": False,
+        "ai_proxy": False,
         "attachment_upload": False,
         "children_create": False,
         "children_list": True,
@@ -276,8 +274,7 @@ def test_models_documents_get_abilities_editor(
     expected_abilities = {
         "accesses_manage": False,
         "accesses_view": False,
-        "ai_transform": is_authenticated,
-        "ai_translate": is_authenticated,
+        "ai_proxy": is_authenticated,
         "attachment_upload": True,
         "children_create": is_authenticated,
         "children_list": True,
@@ -328,8 +325,7 @@ def test_models_documents_get_abilities_owner(django_assert_num_queries):
     expected_abilities = {
         "accesses_manage": True,
         "accesses_view": True,
-        "ai_transform": True,
-        "ai_translate": True,
+        "ai_proxy": True,
         "attachment_upload": True,
         "children_create": True,
         "children_list": True,
@@ -377,8 +373,7 @@ def test_models_documents_get_abilities_administrator(django_assert_num_queries)
     expected_abilities = {
         "accesses_manage": True,
         "accesses_view": True,
-        "ai_transform": True,
-        "ai_translate": True,
+        "ai_proxy": True,
         "attachment_upload": True,
         "children_create": True,
         "children_list": True,
@@ -429,8 +424,7 @@ def test_models_documents_get_abilities_editor_user(django_assert_num_queries):
     expected_abilities = {
         "accesses_manage": False,
         "accesses_view": True,
-        "ai_transform": True,
-        "ai_translate": True,
+        "ai_proxy": True,
         "attachment_upload": True,
         "children_create": True,
         "children_list": True,
@@ -488,8 +482,7 @@ def test_models_documents_get_abilities_reader_user(
         "accesses_view": True,
         # If you get your editor rights from the link role and not your access role
         # You should not access AI if it's restricted to users with specific access
-        "ai_transform": access_from_link and ai_access_setting != "restricted",
-        "ai_translate": access_from_link and ai_access_setting != "restricted",
+        "ai_proxy": access_from_link and ai_access_setting != "restricted",
         "attachment_upload": access_from_link,
         "children_create": access_from_link,
         "children_list": True,
@@ -545,8 +538,7 @@ def test_models_documents_get_abilities_preset_role(django_assert_num_queries):
     assert abilities == {
         "accesses_manage": False,
         "accesses_view": True,
-        "ai_transform": False,
-        "ai_translate": False,
+        "ai_proxy": False,
         "attachment_upload": False,
         "children_create": False,
         "children_list": True,
@@ -592,8 +584,7 @@ def test_models_document_get_abilities_ai_access_authenticated(is_authenticated,
     document = factories.DocumentFactory(link_reach=reach, link_role="editor")
 
     abilities = document.get_abilities(user)
-    assert abilities["ai_transform"] is True
-    assert abilities["ai_translate"] is True
+    assert abilities["ai_proxy"] is True
 
 
 @override_settings(AI_ALLOW_REACH_FROM="authenticated")
@@ -611,8 +602,7 @@ def test_models_document_get_abilities_ai_access_public(is_authenticated, reach)
     document = factories.DocumentFactory(link_reach=reach, link_role="editor")
 
     abilities = document.get_abilities(user)
-    assert abilities["ai_transform"] == is_authenticated
-    assert abilities["ai_translate"] == is_authenticated
+    assert abilities["ai_proxy"] == is_authenticated
 
 
 def test_models_documents_get_versions_slice_pagination(settings):